pyproject.toml
[tool.poetry]
name = "personalshoppingassistant"
version = "0.1.0"
description = ""
authors = ["rex69420 <mridu674385@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src", from = "." },
]

[tool.poetry.dependencies]
python = "3.11.*"
uagents = "^0.9.2"
python-dotenv = "^1.0.0"
bs4 = "^0.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
